Extract CachingList from ExternalSource
