+++ dali/operators/CMakeLists.txt
@@ -22,6 +22,7 @@ add_subdirectory(decoder)
 add_subdirectory(generic)
 add_subdirectory(image)
 add_subdirectory(imgcodec)
+add_subdirectory(input)
 add_subdirectory(math)
 add_subdirectory(random)
 add_subdirectory(reader)
